% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/externVar.R
\name{externVar}
\alias{externVar}
\title{Estimation of external variable latent class models}
\usage{
externVar(
  model,
  method,
  fixed,
  mixture,
  random,
  subject,
  classmb,
  idiag = FALSE,
  nwg = FALSE,
  link = NULL,
  intnodes = NULL,
  epsY = 0.5,
  data,
  B,
  convB = 1e-04,
  convL = 1e-04,
  convG = 1e-04,
  longitudinal,
  maxiter = 100,
  posfix,
  verbose = TRUE,
  varest = "none",
  M = 200,
  nproc = 1
)
}
\arguments{
\item{model}{original model with latent class structure with which external variables
must be modeled.}

\item{method}{character string representing the method to be used :
either "twoStageJoint"}

\item{fixed}{two sided linear formula object for specifying the fixed-effects on external variable}

\item{mixture}{one-sided formula object for the class-specific fixed effects in the linear mixed model
By default, an intercept is included. If no intercept, \code{-1} should be the first term included.}

\item{random}{optional one sided linear formula object for specifying the random-effects on external variable
By default, an intercept is included. If no intercept, \code{-1} should be the first term included.}

\item{subject}{name of the covariate representing the grouping structure.
Even without random effect, either the subject variable used in model or
the subject variable specified as parameter must be present in data.}

\item{classmb}{optional one-sided formula describing the covariates in the
class-membership multinomial logistic model}

\item{idiag}{optional logical for the structure of the variance-covariance
matrix of the random-effects. If \code{FALSE}, a non structured matrix of
variance-covariance is considered (by default).  If \code{TRUE} a diagonal
matrix of variance-covariance is considered.}

\item{nwg}{optional logical indicating if the variance-covariance of the
random-effects is class-specific. If \code{FALSE} the variance-covariance
matrix is common over latent classes (by default). If \code{TRUE} a
class-specific proportional parameter multiplies the variance-covariance
matrix in each class (the proportional parameter in the last latent class
equals 1 to ensure identifiability).}

\item{link}{link optional family of link functions to estimate for the external outcome. Defaults to NULL, corresponding to hlme function}

\item{epsY}{}

\item{data}{: optional data frame containing the variables named used elsewhere.
IMPORTANT : unique individual variables (such as gender or educational level)
MUST only appear once per subject AND at the same visit time.}

\item{B}{: optional specification for the initial values for the parameters.
Two options are allowed: (1) a vector of initial values is entered.
(2) nothing is specified. A preliminary analysis involving the
estimation of a standard model is performed to choose initial
values.}

\item{longitudinal}{optional list of longitudinal models of type hlme,
lcmm or multlcmm used to build mpjlcmm input model for "twoStageJoint" method.
By default, the function will try to retrieve it from input model call.}

\item{maxiter}{optional maximum number of iterations for the Marquardt
iterative algorithm}

\item{posfix}{Optional vector specifying the indices in vector B of the
parameters that should not be estimated. Default to NULL, all external parameters are
estimated.}

\item{verbose}{logical indicating whether information about computation should be
reported. Default to TRUE.}

\item{varest}{character string indicating the method used to account for step one variability
when computing the variance estimation.
either "None", "ParamBoot" or "calc" ("calc" is for "twoStageJoint" method only)}

\item{M}{integer number of parametrical boostrap iterations when varest is "paramBoot". Default to 200.}

\item{nproc}{the number cores for parallel computation. Default to 1 (sequential mode).}
}
\description{
This function allows to model external variables in relationship with previously modeled
latent class structures. The joint likelihood, for both external class
predictors and external outcome of the class.
}
