\name{ForInternalUse}
\alias{mixture}
\alias{factor.names}
\alias{.Ordlcmm}
\alias{.Contlcmm}
\alias{risq_spl}
\alias{risqcum_spl}
\alias{.plotresid}
\alias{.plotpostprob}
\alias{.plotfit}
\alias{.plotlinkfuntion}
\alias{.plotlinkfunctionmult}
\alias{.plotbaselinerisk}
\alias{.plotsurvival}



\title{For internal use only ... }

\description{For internal use only ... }
