% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinal_score_data.R
\docType{data}
\name{ordinal_score_data}
\alias{ordinal_score_data}
\title{Simulated dataset for the location-scale model in multlcmm function}
\format{
A data frame with 3900 observations over 300 different subjects and
3 variables.  \describe{ \item{i}{subject identification number}
\item{score}{longitudinal ordinal outcome with 5 levels}
\item{time}{time of measurement}
\item{treatment}{binary covariate}}
}
\description{
The data were simulated from a latent process location-scale mixed model with 5-levels
ordinal outcome.  Repeated data of the longitudinal outcome (score) were
simulated for a total of 300 subjects in fixed time points. Every subject
has measurements in equal time intervals of 0.23 month, i.e. around one week. 
The follow-up is 3 months for all individuals.  The outcome was
predicted by the treatment covariate and splines for time variable.  }
\seealso{
\code{\link{multlcmm}}}
\keyword{datasets}

