% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictClass.R
\name{predictClass}
\alias{predictClass}
\title{Posterior classification and class-membership probabilities}
\usage{
predictClass(model, newdata, subject = NULL)
}
\arguments{
\item{model}{an object inheriting from class \code{lcmm}, 
\code{Jointlcmm} or \code{multlcmm} representing a general latent class
mixed model.}

\item{newdata}{data frame containing the data from which predictions are computed. The data frame should include at least all the covariates listed in model$Xnames2. Names in the data frame should be exactly model$Xnames2 that are the names of covariates specified in lcmm, hlme, Jointlcmm or multlcmm calls.}

\item{subject}{choice of subject, by default: NULL, meaning the subject identifier is the same as in the model.}
}
\value{
dataframe containing the subject identifier, the predicted class and the posterior probabilities.
}
\description{
This function provides a table of posterior classification and posterior individual class-membership probabilities with new data
}
\examples{
paquid$normMMSE <- normMMSE(paquid$MMSE)
paquid$age65 <- (paquid$age - 65)/10
m2b <- hlme(normMMSE ~ age65+I(age65^2)+CEP, random =~ age65+I(age65^2), subject = 'ID',data = paquid, ng = 2, mixture =~ age65+I(age65^2), B = c(0, 60, 40, 0, -4, 0, -10, 10, 212.869397, -216.421323,456.229910, 55.713775, -145.715516, 59.351000, 10.072221))
predictClass(m2b, newdata)
}
\author{
Cecile Proust-Lima, Viviane Philipps, Sasha Cuau
}
