% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate.lcmm}
\alias{simulate.lcmm}
\title{Data simulation according to a latent class mixed model}
\usage{
\method{simulate}{lcmm}(
  object,
  nsim,
  seed,
  times,
  tname = NULL,
  n,
  Xbin = NULL,
  Xcont = NULL,
  entry = NULL,
  dropout = NULL,
  pMCAR = 0,
  ...
)
}
\arguments{
\item{object}{an object of class \code{hlme}, \code{lcmm}, \code{multlcmm} or
\code{Jointlcmm}}

\item{nsim}{not used (for compatibility with stats::simulate). The function simulates only one sample}

\item{seed}{the random seed}

\item{times}{either a data frame with 2 columns containing IDs and measurement times, or a vector of length 4 specifying the minimal and maximum measurement times, the spacing between 2 consecutive visits and the margin around this spacing}

\item{tname}{the name of the variable represneting the measurement times in the object.
Default to the second column's name of times if it is a data frame, and to object$var.time otherwise.}

\item{n}{number of subject to simulate. Required only if times is not a data frame}

\item{Xbin}{an optional named list giving the probabilities of the binary
covariates to simulate. The list's names should match the binary covariate's names
used in \code{object}.}

\item{Xcont}{an optional named list giving the mean and standard deviation
of the Gaussian covariates to simulate. The list's names should match the
continuous covariate's names used in \code{object}.}

\item{entry}{expression to simulate a subject's entry time}

\item{dropout}{expression to simulate a subject's time to dropout}

\item{pMCAR}{optional numeric giving an observation's probability to be missing}

\item{\dots}{additionnal options. None is used yet.}
}
\value{
a data frame with one line per observation and one column per variable.
}
\description{
This function simulates a sample according to a model estimated with \code{hlme},
\code{lcmm}, \code{multlcmm} or \code{Jointlcmm} functions.
}
\author{
Viviane Philipps and Cecile Proust-Lima
}
