% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryplot.R
\name{summaryplot}
\alias{summaryplot}
\title{Summary of models}
\usage{
summaryplot(
  m1,
  ...,
  which,
  width = length(which),
  height = 1,
  xaxis = "G",
  display = FALSE
)
}
\arguments{
\item{m1}{an object of class \code{hlme}, \code{lcmm}, \code{multlcmm} or
\code{Jointlcmm}}

\item{\dots}{further arguments, in particular other objects of class
\code{hlme}, \code{lcmm}, \code{multlcmm} or \code{Jointlcmm}}

\item{which}{character vector indicating which results should be returned.
Possible values are "loglik", "conv", "npm", "AIC", "BIC", "SABIC",
"entropy", "ICL", "\%class".}

\item{xaxis}{the abscissa of the plot, default to "G}

\item{display}{display of the table. By default, False.}
}
\description{
This function provides a plot summarizing the results of different models
fitted by \code{hlme}, \code{lcmm}, \code{multlcmm} or \code{Jointlcmm}.
}
\details{
Can be reported the usual criteria used to assess the fit and the clustering
 of the data:
 - maximum log-likelihood L (the higher the better)
 - number of parameters P, number of classes G, convergence criterion (1 = converged)
 - AIC (the lower the better) computed as -2L+2P 
 - BIC (the lower the better) computed as -2L+ P log(N) where N is the number of subjects
 - SABIC (the lower the better) computed as 
 - Entropy (the closer to one the better) computed as 1-sum[pi_ig*log(pi_ig)]/(N*log(G))
   where pi_ig is the posterior probability that subject i belongs to class g
 - ICL (the lower the better) computed as BIC -2*sum[c_ig*log(pi_ig)]
   where c_ig is the posterior class membership
 - %Class computed as the proportion of each class based on c_ig
}
\examples{
  
}
\seealso{
\code{\link{summary}}, \code{\link{summarytable}}
}
\author{
Cecile Proust-Lima, Viviane Philipps, Sasha Cuau
}
