% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.mpjlcmm.R
\name{update.mpjlcmm}
\alias{update.mpjlcmm}
\title{Update the longitudinal submodels}
\usage{
\method{update}{mpjlcmm}(object, ...)
}
\arguments{
\item{object}{an estimated mpjlcmm model}

\item{\dots}{not used}
}
\value{
A list of hlme/lcmm/multlcmm models. The models appear in the same order as
specified in the call to the mpjlcmm function.
}
\description{
This function updates the longitudinal submodels of a mpjlcmm object by injecting
the estimated parameters and their variances in each hlme/lcmm/multlcmm model used
to define the multi-process joint model.
The same (uni-process) models as specified in the mpjlcmm call are returned, with
updated outputs for best, V, conv, cholesky, pred, predRE, predRE_Y, pprob.
All postfit functions (plots, predictions, ...) can then be called on the updated hlme/lcmm/multlcmm models.
See mpjlcmm's help page for examples.
}
