% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarCov.R
\name{VarCov}
\alias{VarCov}
\title{Variance-covariance of the estimates}
\usage{
VarCov(x)
}
\arguments{
\item{x}{an object of class \code{hlme}, \code{lcmm}, \code{multlcmm},
\code{Jointlcmm} or \code{mpjlcmm}}
}
\value{
a matrix containing the variance-covariance of the estimates. For
the parameters of the matrix of variance-covariance of the random effects,
the Cholesky transformed parameters are considered so that VarCov provides
the covariance matrix of function \code{estimates} with cholesky=TRUE.
}
\description{
This function provides the variance-covariance matrix of the estimates. vcov
is an alias for it.
}
\seealso{
\code{\link{estimates}}
}
\author{
Cecile Proust-Lima, Viviane Philipps
}
