% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarCovRE.hlme.R
\name{VarCovRE}
\alias{VarCovRE}
\alias{VarCovRE.hlme}
\alias{VarCovRE.lcmm}
\alias{VarCovRE.Jointlcmm}
\alias{VarCovRE.multlcmm}
\title{Estimates, standard errors and Wald test for the parameters of the
variance-covariance matrix of the random effects.}
\usage{
VarCovRE(Mod)
}
\arguments{
\item{Mod}{an object of class \code{hlme}, \code{lcmm}, \code{multlcmm} or
\code{Jointlcmm}}
}
\value{
a matrix containing the estimates of the parameters of the
variance-covariance matrix of the random effects, their standard errors,
and, for the covariance parameters, the Wald statistic and the associated
p-value.
}
\description{
Fromm the Cholesky transformed parameters, this function provides estimates,
standard errors and Wald test for the parameters of the variance-covariance
matrix of the random effects.
}
\author{
Cecile Proust-Lima, Lionelle Nkam and Viviane Philipps
}
