% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.hlme.R
\name{estimates}
\alias{estimates}
\alias{estimates.hlme}
\alias{estimates.lcmm}
\alias{estimates.Jointlcmm}
\alias{estimates.multlcmm}
\title{Maximum likelihood estimates}
\usage{
estimates(x, cholesky = TRUE)
}
\arguments{
\item{x}{an object of class \code{hlme}, \code{lcmm}, \code{multlcmm} or
\code{Jointlcmm}}

\item{cholesky}{optional logical indicating if the parameters of
variance-covariance of the random effets should be displayed instead of
their cholesky transformations used in the estimation process.}
}
\value{
a vector with all estimates of the model.
}
\description{
This function provides the vector of maximum likelihood estimates of a model
estimated with \code{hlme}, \code{lcmm}, \code{multlcmm} or
\code{Jointlcmm}.
}
\seealso{
\code{\link{VarCov}}, \code{\link{hlme}}, \code{\link{lcmm}},
\code{\link{multlcmm}}, \code{\link{Jointlcmm}}
}
\author{
Cecile Proust-Lima, Viviane Philipps
}
