\name{ForInternalUse}
\alias{mixture}
\alias{factor.names}
\alias{.Ordlcmm}
\alias{.Contlcmm}
\alias{risq_spl}
\alias{risqcum_spl}
\alias{.plotresid}
\alias{.plotpostprob}
\alias{.plotfit}
\alias{.plotlinkfuntion}
\alias{.plotlinkfunctionmult}
\alias{.plotbaselinerisk}
\alias{.plotsurvival}
\alias{C_hetmixlin}
\alias{C_hetmixcont}
\alias{C_hetmixord}
\alias{C_hetmixcontmult}
\alias{C_jointhet}
\alias{C_predictcont}
\alias{C_predictmult}
\alias{C_postprob2}
\alias{C_cvpl}
\alias{C_calculustransfo}




\title{For internal use only ... }

\description{For internal use only ... }
