\name{lcmm-package}
\alias{lcmm-package}
\docType{package}
\title{
Estimation of Latent Class Mixed Models and Joint Latent Class Mixed Models for quantitative, bounded quantitative, discrete and ordinal longitudinal outcomes and a right-censored left-truncated time-to-event.
}
\description{
This package provides functions for the estimation of latent class mixed models (LCMM) and joint latent class mixed models using a maximum likelihood method.
}
\details{
\tabular{ll}{
Package: \tab lcmm \cr
Type: \tab Package \cr
Version: \tab 1.4-3 \cr
Date: \tab 2011-01-13 \cr
License: \tab GPL (>=2.0) \cr
LazyLoad: \tab yes \cr
}
The package includes for the moment the estimation of latent class mixed models for Gaussian longitudinal outcomes using \code{hlme} function, and for other quantitative, bounded quantitative and discrete longitudinal outcomes using \code{lcmm} function, and joint latent class mixed model for Gaussian longitudinal outcome and a time-to-event using \code{Jointlcmm} function. 

Please report to the maintainer any bug or comment regarding the package for future updates. 
}
\author{
Cecile Proust-Lima, Amadou Diakite and Benoit Liquet

Maintainer: <cecile.proust@isped.u-bordeaux2.fr>
}
\references{
Lin, Turnbull, McCulloch and Slate (2002). Latent class models for
joint analysis of longitudinal biomarker and event process data: application to longitudinal
prostate-specific antigen readings and prostate cancer. Journal of the American Statistical
Association 97, 53-65.

Muthen and Shedden (1999). Finite mixture modeling with mixture outcomes using the EM algorithm. Biometrics 55, 463-9

Proust and Jacqmin-Gadda (2005). Estimation of linear mixed models with a mixture of distribution for the random-effects. Comput Methods Programs Biomed 78:165-73

Proust, Jacqmin-Gadda, Taylor, Ganiayre, and Commenges (2006). A
nonlinear model with latent process for cognitive evolution using multivariate longitudinal
data. Biometrics 62, 1014-24.

Proust-Lima, Dartigues, and Jacqmin-Gadda (2011). Misuse of the linear mixed
model when evaluating risk factors of cognitive decline. Amer J Epidemiol - in press

Proust-Lima and Taylor (2009). Development and validation of a dynamic prognostic
tool for prostate cancer recurrence using repeated measures of post-treatment PSA: a joint
modelling approach. Biostatistics 10, 535-49.

Verbeke and Lesaffre (1996). A linear mixed-effects model with heterogeneity in
the random-effects population. Journal of the American Statistical Association 91, 217-21
}
\keyword{package}

