% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.lcmm.R
\name{print.lcmm}
\alias{print.lcmm}
\alias{print.hlme}
\alias{print.Jointlcmm}
\alias{print.multlcmm}
\alias{print.epoce}
\alias{print.Diffepoce}
\title{Brief summary of a \code{hlme}, \code{lcmm},
\code{Jointlcmm},\code{multlcmm}, \code{epoce} or \code{Diffepoce} objects}
\usage{
\method{print}{lcmm}(x, ...)
}
\arguments{
\item{x}{an object inheriting from classes \code{hlme}, \code{lcmm} for
fitted latent class mixed-effects, or class \code{Jointlcmm}for a Joint
latent class mixed model or \code{epoce} for predictive accuracy
computations.}

\item{\dots}{further arguments to be passed to or from other methods.  They
are ignored in this function.}
}
\description{
The function provides a brief summary of \code{hlme},
\code{lcmm},\code{multlcmm} or \code{Jointlcmm} estimations, and
\code{epoce} or \code{Diffepoce} computations.
}
\seealso{
\code{\link{hlme}}, \code{\link{lcmm}}, \code{\link{Jointlcmm}},
\code{epoce}, \code{Diffepoce}
}
\author{
Cecile Proust-Lima, Viviane Philipps, Amadou Diakite and Benoit
Liquet
}
\keyword{print}
