% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarytable.R
\name{summarytable}
\alias{summarytable}
\title{Summary of models}
\usage{
summarytable(m1, ..., which = c("G", "loglik", "npm", "BIC", "\%class"))
}
\arguments{
\item{m1}{an object of class \code{hlme}, \code{lcmm}, \code{multlcmm} or
\code{Jointlcmm}}

\item{\dots}{further arguments, in particular other objects of class
\code{hlme}, \code{lcmm}, \code{multlcmm} or \code{Jointlcmm}}

\item{which}{character vector indicating which results should be returned.
Possible values are "G", "loglik", "conv", "npm", "AIC", "BIC", "SABIC",
"entropy", "\%class".}
}
\value{
a matrix giving for each model the values of the requested indexes.
By default, the number a latent classes, the
log-likelihood, the number of parameters, the BIC and the posterior
probability of the latent classes.
}
\description{
This function provides a table summarizing the results of different models
fitted by \code{hlme}, \code{lcmm}, \code{multlcmm} or \code{Jointlcmm}.
}
\seealso{
\code{\link{summary}}, \code{\link{hlme}}, \code{\link{lcmm}},
\code{\link{multlcmm}}, \code{\link{Jointlcmm}}
}
\author{
Cecile Proust-Lima, Viviane Philipps
}
